build: do not check for the backend in when using in-tree backends
Right now we are trying to check for in-tree backends, which obviously
won't work.

We only have this check to provide nicer error messages to users, I
don't see the need to provide the same "nicer" error messages when using
in-tree backends as it is quite niche.

In these cases, pep517 will raise an error and we will print it.

Signed-off-by: Filipe Laíns <lains@archlinux.org>