Remove check for the existence of build-backend
This has been moved out of the ProjectBuilder constructor, since the
ProjectBuilder is constructed before the isolated build environment is
constructed and not executed IN the isolated environment, this check is
not relevant.

In its place, we add a wrapper around the exception that pep517 raises
when the backend cannot be imported.

Fixes GH-98.