env: pass --use-pep517 to force build isolation
In #231 we discovered pip is not building some dependencies in
isolation, it is using the build environment, which can become an issue
since we remove setuptools from the environment.

As suggested by @layday, this patch passes --use-pep517 to pip so that
it builds in isolation, as recomended by PEP 517.

Fixes #231

Signed-off-by: Filipe Laíns <lains@riseup.net>