Fix sphinx syntax @ `--config-setting` doc
RST treats single backticks as `:any:` role that tries
to automatically match something with that name. This
generates a warning because there's nothing to be found.
But running Sphinx in nitpicky mode turns the warning
into an error.

There's two ways of fixing this: either use double backticks
for inline code or drop them and maybe switch to double-
quotes. This change does the former.