🎨🧪 Modularize GHA workflow through reuse
This change slightly refactors parts of the GitHub Actions CI/CD
workflow definitions to make it possible to rely on the main check
gate in branch protection and declutter the generic workflow.

It reproduces the original efforts of skipping the unnecessary checks
depending on the files changed via dynamically determined variables.