fix: stop injecting `wheel` as a build dep fallback (#716)
PEP 517 doesn't mandate depending on `wheel` when a `__legacy__`
setuptools fallback is used. Historically, it used to be assumed as
necessary, but later it turned out to be wrong. The reason is that
`setuptools`' `get_requires_for_build_wheel()` hook already injects
this dependency when building wheels is requested [[1]].
It also used to have this hint in the docs, but it was corrected
earlier [[2]].

This patch removes `wheel` from said `requires` list fallback.

[1]: https://github.com/pypa/setuptools/blob/v40.8.0/setuptools/build_meta.py#L130
[2]: https://github.com/pypa/setuptools/pull/3056