docs: Add method of installation via conda
On conda-forge, this package is called "python-build", not "build".
However, this information is only given in README.md and not in
docs/installation.rst.  Users are likely to read the documentation on
https://build.pypa.io/ only, and they are likely to think that the name
of the package on conda-forge is also "build", which will make them get
into trouble.  Fix this problem here.