fix: Fix `ModuleNotFoundError` when `pip` is not installed (#898)
* fix: Fix `ModuleNotFoundError` when `pip` is not installed

Fix the exception raised if `pip` is not installed at all:

```
[...]
  File "/usr/lib/python3.14/site-packages/build/env.py", line 168, in _has_valid_outer_pip
    if importlib.util.find_spec('pip._vendor') is None:
       ~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^
  File "<frozen importlib.util>", line 90, in find_spec
ModuleNotFoundError: No module named 'pip'
```

This is a regression from #861 where I did not predict that
`find_spec()` will actually raise when the parent package
is not available. To fix it, just reorder the conditions to make sure
that `pip` is installed first.

* Update src/build/env.py

---------

Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>